+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2026-02-14 03:57:54
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: 9b308869be78609019fd23636535b4f9881bc9b5
+
+## [ForwardWidgets-2kuai] - 2026-02-14 03:57:54
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: edf5dc0e8b22380f9beb3ab70606aa4b3d0c4420
+
+## [ForwardWidgets-pack1r] - 2026-02-14 03:57:54
+- 仓库: https://github.com/pack1r/ForwardWidgets
+- 分支: main
+- 新增: b89e7e6abbbc02434211fd02fc0d46b2b8cafeae
+
+
 ## [ForwardWidgets-huangxd] - 2026-02-13 04:07:32
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main